Hand an orchestrator's prompt over intact, and make the review mirror fillable without the desktop
Every prompt-carrying orchestrator launch died before its session started, so
the Investigate flow and the post-rebuild auto-resume both lost their task. Two
independent defects, either one fatal.

The prompt was appended after --mcp-config, which takes multiple values, so the
CLI consumed the prompt as another config path and resolved it against the
working directory ("MCP config file not found: <cwd>/Resumed after a
rebuild..."; ENAMETOOLONG for a long one). Option parsing now ends with -- before
the positional prompt.

The prompt was also spliced into the host shell's command line in double quotes,
which neutralize nothing: a prompt carrying ordinary markdown code spans had them
executed on the operator's machine, deleted from what the harness received, and —
with a backslash or an odd backtick — collapsed the argument boundaries around
them. Both the prompt and the config path now go through the shell quoting
erun-ui already owns for exactly this.

The launch tests now let a real POSIX shell parse the composed command and report
the argv it produces, so the guarantee is measured rather than asserted.

Alongside that, the workspace-sync pass moves from erun-ui to erun-common and
gains the two transports it lacked. It was reachable only from the desktop's
poller, so an orchestrator whose mirror was empty had no erun-native way to
repair its own read surface and paid for hand-rolled workarounds instead. The
desktop poller now calls the shared pass rather than owning it, `erun sshd sync`
runs one pass, and the `workspace_sync` MCP tool runs the same one. That tool is
served by the host-side proxy rather than relayed to the edge, because the edge
runs in the pod and the mirror is on the host.

All three transports resolve through one shared resolution, so they agree on what
a pass addresses and refuse by naming which precondition failed. Preview reads the
same listings and stops before every write, so a dry run reports real counts.

The sync pass now resolves its ssh/tar children through the ERUN_<NAME>_BIN seam
every other erun child uses; without it the pass silently reached for whatever the
host had installed and could not be exercised.

The workspace-sync spec described a local-to-pod mirror into /home/erun/sync with
a sweeper goroutine and scp. None of that exists in the code, and the real
direction is the opposite; the page now documents the pass as implemented.

Closes #936
Closes #938
