+++ erun-docs/docs/cli/sshd.md
@@ -10,6 +10,7 @@ Enable SSH access to a remote environment — the prerequisite for attaching an 
 
 ```
 erun sshd init [TENANT] [ENVIRONMENT] [flags]
+erun sshd sync [TENANT] [ENVIRONMENT] [flags]
 ```
 
 ## What it does
@@ -36,6 +37,21 @@ erun sshd init my-tenant rihards-dev --dry-run
 erun sshd init my-tenant rihards-dev
 erun open my-tenant rihards-dev --vscode   # now possible
 ```
+
+## `erun sshd sync`
+
+Runs one workspace-sync pass for a `remote-agent` environment: mirrors the pod's git-visible worktree into the host review directory, deletes what the pod no longer has, and delivers the pod's cross-built artifacts into the mirror's `.erun-outputs/`. See the [workspace sync spec](/agent-reference/workspace-sync-spec) for the pass itself.
+
+The desktop runs the same pass on a poller. This command exists so the pass is reachable without the desktop — an orchestrator whose mirror is empty or stale refreshes it in one call instead of working around it.
+
+```bash
+erun sshd sync my-tenant rihards-dev --dry-run   # counts only, mirror untouched
+erun sshd sync my-tenant rihards-dev
+```
+
+`--dry-run` resolves the pass, traces the pod path and the host path it maps to, and reports the counts a real pass would change without creating, fetching, or deleting anything.
+
+It refuses, naming which precondition failed, when the environment has no pod worktree (not a `remote-agent` env), has workspace sync disabled, has no configured local path, or when its SSH channel is not up.
 
 ## Error behaviour
