+++ erun-integration/testdata/mcp/proxy_real_run_recovers_when_the_edge_forgets_its_session.txt
@@ -1,5 +1,5 @@
 {"jsonrpc":"2.0","id":1,"result":{"protocolVersion":"2025-06-18","capabilities":{"tools":{}}}}
-{"jsonrpc":"2.0","id":2,"result":{"tools":[{"name":"version","description":"Report the runtime's erun version"}]}}
+{"jsonrpc":"2.0","id":2,"result":{"tools":[{"name":"version","description":"Report the runtime's erun version"},{"name":"workspace_sync","description":"Run one workspace-sync pass for this environment: mirror the pod's git-visible worktree into the host review directory, delete what the pod no longer has, and deliver the pod's cross-built artifacts. Runs on the host, where the mirror lives. Set preview to report what a pass would change without touching the mirror. Refuses, naming which, when the environment is not a remote-agent env, has workspace sync disabled, has no configured local path, or its SSH channel is down.","inputSchema":{"type":"object","properties":{"preview":{"type":"boolean","description":"when true, report what one pass would change without changing it"}}}}]}}
 audit: erun mcp proxy --environment dev --tenant team
 mcp: team/dev edge resolved to http://<LOOPBACK>:26100/mcp
 mcp proxy: the MCP edge at http://<LOOPBACK>:26100/mcp had forgotten its session; re-initialized and retried
